+++ hal/vulkan/vk/signatures.go
@@ -221,6 +221,9 @@ var (
 
 	// void(handle, handle, u32, u32, handle, u64, u64, u32) - vkCmdCopyQueryPoolResults
 	SigVoidCmdCopyQueryPoolResults types.CallInterface
+
+	// VkResult(handle, ptr, u64) - vkWaitSemaphores
+	SigResultHandlePtrU64 types.CallInterface
 )
 
 // InitSignatures prepares all CallInterface templates.
@@ -705,6 +708,13 @@ func InitSignatures() error {
 	// void(handle, handle, u32, u32, handle, u64, u64, u32) - vkCmdCopyQueryPoolResults
 	err = ffi.PrepareCallInterface(&SigVoidCmdCopyQueryPoolResults, types.DefaultCall, voidRet,
 		[]*types.TypeDescriptor{u64, u64, u32, u32, u64, u64, u64, u32})
+	if err != nil {
+		return err
+	}
+
+	// VkResult(handle, ptr, u64) - vkWaitSemaphores
+	err = ffi.PrepareCallInterface(&SigResultHandlePtrU64, types.DefaultCall, resultRet,
+		[]*types.TypeDescriptor{u64, ptr, u64})
 	if err != nil {
 		return err
 	}